requirements.txt
aiohttp~=3.7.1
asyncpg~=0.21.0
discord.py~=1.5.1